Add Content
Added the ability to have content in the menu dropdown (i.e. the so-called "mega menu").  Also added "flipping": The content will always be on the "inside" (i.e. if the menu has to display to the left of its associated nav item the content will be on the left, and if it has to display to the right of its associated nav item the content will be on the right).