.open
The JavaScript code which adds hover support now adds/removes the "hover" class from the element with the "dropdown" class.  This undermines the fadeIn/fadeOut since Bootstrap CSS shows/hides the descendant elements automatically in the presence/absence of the "open" class and therefore those calls have been removed.  It may be possible to work around this if the animations are still desired.