Intelligent Flipping & .no-fit
Dropdowns now hang off to the left (i.e. they're attached at the right side to their parent) when they're on the right side of the screen and to the right (i.e. they're attached to the left side of their parent) when they're on the left side of the screen.  This feels more natural/more symmetric.

Also: If a dropdown does not fit to the left or to the right it will have the "no-fit" class added to it.